src/requirements.txt
Flask==0.11.1
Flask-Login==0.3.2
Flask-SQLAlchemy==2.1
requests==2.11.1
passlib==1.6.5
PYMySQL3==0.5
Jinja2==2.8
Werkzeug